+++ verl/experimental/agent_loop/load_balance.py
@@ -11,224 +11,150 @@
 # WITHOUT WARRANTIES OR ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-"""Pluggable load-balance strategies for :class:`GlobalRequestLoadBalancer`."""
+"""
+Ray global load balancers:
+- Concrete actors use @ray.remote.
+- load_balancer_actor_class maps rollout.load_balance_sticky_mode to the actor class.
+"""
 
 from __future__ import annotations
 
-import logging
-import random
-import threading
 from abc import ABC, abstractmethod
 from typing import Any
 
-from verl.experimental.agent_loop.prometheus_metrics import (
-    build_metrics_url,
-    fetch_prometheus_text,
-    parse_prometheus_metric_value,
-)
+import ray
+from cachetools import LRUCache
 
-logger = logging.getLogger(__name__)
+from verl.experimental.agent_loop.load_balance_strategy import create_load_balance_strategy
+from verl.workers.config import RolloutConfig
 
-_LOAD_BALANCE_STRATEGY_CLASSES: dict[str, type[LoadBalanceStrategy]] = {}
+DEFAULT_ROUTING_CACHE_SIZE = 10000
 
-
-def _register(name: str, cls: type[LoadBalanceStrategy]) -> None:
-    if name in _LOAD_BALANCE_STRATEGY_CLASSES:
-        raise ValueError(f"Load balance strategy {name!r} is already registered")
-    _LOAD_BALANCE_STRATEGY_CLASSES[name] = cls
+_LOAD_BALANCER_ACTOR_CLASSES: dict[str, Any] = {}
 
 
-def register(name: str):
-    """Decorator: @register("my_strategy") on a :class:LoadBalanceStrategy subclass (like agent loop @register)."""
-
-    def decorator(cls: type[LoadBalanceStrategy]) -> type[LoadBalanceStrategy]:
-        _register(name, cls)
-        return cls
-
-    return decorator
-
-
-def register_load_balance_strategy(name: str, cls: type[LoadBalanceStrategy]) -> None:
-    """Register a strategy class without decorator (e.g. plugins). Same as ``@register(name)`` on ``cls``."""
-    _register(name, cls)
-
-
-def create_load_balance_strategy(name: str, server_actor_ids: list[str], **kwargs: Any) -> LoadBalanceStrategy:
-    """Instantiate a registered strategy by name."""
-    cls = _LOAD_BALANCE_STRATEGY_CLASSES.get(name)
+def load_balancer_actor_class(rollout_config: RolloutConfig) -> Any:
+    """Resolve the Ray actor class from rollout_config.load_balance_sticky_mode."""
+    mode = rollout_config.load_balance_sticky_mode
+    cls = _LOAD_BALANCER_ACTOR_CLASSES.get(mode)
     if cls is None:
         raise ValueError(
-            f"Unknown load_balance_strategy {name!r}. Registered: {sorted(_LOAD_BALANCE_STRATEGY_CLASSES)}"
+            f"Unknown load_balance_sticky_mode {mode!r}. Registered: {sorted(_LOAD_BALANCER_ACTOR_CLASSES)}"
         )
-    return cls(server_actor_ids, **kwargs)
+    return cls
 
 
-class LoadBalanceStrategy(ABC):
-    """Stateless or stateful strategy object used for new routing decisions (sticky handled elsewhere).
+def register_load_balancer_class(name: str, remote_actor_cls: Any) -> None:
+    """Register a Ray actor class without decorator (remote_actor_cls must be ray.remote-wrapped)."""
+    if name in _LOAD_BALANCER_ACTOR_CLASSES:
+        raise ValueError(f"Load balancer actor {name!r} is already registered")
+    _LOAD_BALANCER_ACTOR_CLASSES[name] = remote_actor_cls
 
-    Subclasses must implement pick_server and accept
-    __init__(self, server_actor_ids: list[str], **kwargs)
-    """
+
+class GlobalRequestLoadBalancer(ABC):
+    """Abstract base (not a Ray actor). Subclasses are @ray.remote and listed in _LOAD_BALANCER_ACTOR_CLASSES."""
+
+    @staticmethod
+    def init_bundle_from_rollout(rollout_config: RolloutConfig) -> dict[str, Any]:
+        """Map RolloutConfig load-balance fields to constructor fields (for tests and internal use)."""
+        kv = rollout_config.kv_cache_metrics
+        strategy_init_kwargs: dict[str, Any] = {}
+        if rollout_config.load_balance_weights is not None:
+            strategy_init_kwargs["weights"] = list(rollout_config.load_balance_weights)
+        if rollout_config.load_balance_random_seed is not None:
+            strategy_init_kwargs["random_seed"] = rollout_config.load_balance_random_seed
+        if rollout_config.load_balance_strategy == "least_kv_cache":
+            strategy_init_kwargs["metric_name"] = kv.metric_name
+            strategy_init_kwargs["metrics_path"] = kv.metrics_path
+            strategy_init_kwargs["refresh_interval_s"] = kv.refresh_interval_s
+            strategy_init_kwargs["fetch_timeout_s"] = kv.fetch_timeout_s
+        return {
+            "max_cache_size": DEFAULT_ROUTING_CACHE_SIZE,
+            "load_balance_strategy": rollout_config.load_balance_strategy,
+            "strategy_init_kwargs": strategy_init_kwargs,
+        }
+
+    def _init_from_rollout(self, server_actor_ids: list[str], rollout_config: RolloutConfig) -> None:
+        if not server_actor_ids:
+            raise ValueError("server_actor_ids must be non-empty")
+        bundle = GlobalRequestLoadBalancer.init_bundle_from_rollout(rollout_config)
+        load_balance_strategy = bundle["load_balance_strategy"]
+        if load_balance_strategy == "weighted_rr" and rollout_config.load_balance_weights is not None:
+            nw = len(rollout_config.load_balance_weights)
+            ns = len(server_actor_ids)
+            if nw != ns:
+                raise ValueError(f"load_balance_weights length {nw} must match number of rollout replicas {ns}")
+        self._routing_cache_size = bundle["max_cache_size"]
+        self._server_actor_ids = list(server_actor_ids)
+        sk = bundle["strategy_init_kwargs"] or {}
+        self.strategy = create_load_balance_strategy(
+            load_balance_strategy,
+            server_actor_ids=self._server_actor_ids,
+            **sk,
+        )
 
     @abstractmethod
-    def pick_server(
-        self,
-        server_ids: list[str],
-        inflight: dict[str, int],
-    ) -> str:
-        """Choose server_ids member."""
+    def acquire_server(self, request_id: str, request_group_id: str | None = None) -> str:
+        """Acquire a server. request_group_id ties GRPO / rollout-n repeats to one replica (see trajectory)."""
+
+    def release_server(self, server_id: str) -> None:
+        """Release a server after a request completes, decrementing strategy in-flight counts."""
+        inflight = getattr(self.strategy, "_inflight", None)
+        if inflight is None:
+            raise ValueError("load balance strategy does not expose _inflight; cannot release_server")
+        if server_id not in inflight:
+            raise ValueError(f"Invalid server_id for release: {server_id}")
+        if inflight[server_id] <= 0:
+            raise ValueError(f"Release called with no inflight requests on server {server_id}")
+        inflight[server_id] -= 1
+
+    def update_inflight(self, server_id: str, delta: int) -> None:
+        inflight = getattr(self.strategy, "_inflight", None)
+        if inflight is not None and server_id in inflight:
+            inflight[server_id] += delta
 
 
-@register("least_requests")
-class LeastRequestsStrategy(LoadBalanceStrategy):
-    def __init__(self, server_actor_ids: list[str], **kwargs: Any):
-        pass
+@ray.remote
+class RequestStickyLoadBalancer(GlobalRequestLoadBalancer):
+    """Sticky by request_id only; request_group_id is ignored."""
 
-    def pick_server(
-        self,
-        server_ids: list[str],
-        inflight: dict[str, int],
-    ) -> str:
-        # O(n): one pass for min inflight; ties broken by lexicographically smallest sid (same as sorted(server_ids)).
-        if not server_ids:
-            raise ValueError("server_ids must be non-empty")
-        server_id = min(server_ids, key=lambda sid: (inflight[sid], sid))
+    def __init__(self, server_actor_ids: list[str], rollout_config: RolloutConfig):
+        self._init_from_rollout(server_actor_ids, rollout_config)
+        self._request_id_to_server: LRUCache = LRUCache(maxsize=self._routing_cache_size)
+
+    def acquire_server(self, request_id: str, request_group_id: str | None = None) -> str:
+        if request_id in self._request_id_to_server:
+            server_id = self._request_id_to_server[request_id]
+            self.update_inflight(server_id, 1)
+            return server_id
+        server_id = self.strategy.pick_server(list(self._server_actor_ids))
+        self._request_id_to_server[request_id] = server_id
+        self.update_inflight(server_id, 1)
         return server_id
 
 
-@register("least_kv_cache")
-class LeastKVCacheStrategy(LoadBalanceStrategy):
-    """Prefer lowest KV usage from periodic HTTP /metrics scrape; else fall back to least in-flight.
-
-    When metric_name is set, a daemon thread scrapes each replica on refresh_interval_s and
-    pick_server uses that internal state. When metric_name is unset, no HTTP is performed and
-    all KV readings are treated as unknown (same as least in-flight, with lexicographic sid tie-break).
+@ray.remote
+class GroupStickyLoadBalancer(GlobalRequestLoadBalancer):
+    """
+    GRPO-style group sticky:
+    - Same request_group_id (e.g. step+sample) pins rollout-n repeats.
     """
 
-    def __init__(self, server_actor_ids: list[str], **kwargs: Any):
-        self._server_actor_ids = list(server_actor_ids)
-        self._metric_name: str | None = kwargs.pop("metric_name", None)
-        self._metrics_path: str = kwargs.pop("metrics_path", "/metrics")
-        self._refresh_interval_s: float = float(kwargs.pop("refresh_interval_s", 2.0))
-        self._fetch_timeout_s: float = float(kwargs.pop("fetch_timeout_s", 2.0))
+    def __init__(self, server_actor_ids: list[str], rollout_config: RolloutConfig):
+        self._init_from_rollout(server_actor_ids, rollout_config)
+        self._request_group_to_server: LRUCache[str, str] = LRUCache(maxsize=self._routing_cache_size)
 
-        self._kv_usage: dict[str, float | None] = {sid: None for sid in self._server_actor_ids}
-        self._lock = threading.Lock()
-        self._metrics_stop = threading.Event()
-        self._metrics_thread: threading.Thread | None = None
-        self._metrics_fail_warned: set[str] = set()
-
-        if self._metric_name:
-            self._metrics_thread = threading.Thread(
-                target=self._metrics_background_loop,
-                name="verl-lb-least-kv-cache",
-                daemon=True,
-            )
-            self._metrics_thread.start()
-
-    def _metrics_background_loop(self) -> None:
-        while not self._metrics_stop.is_set():
-            self._refresh_metrics_blocking()
-            if self._metrics_stop.wait(timeout=self._refresh_interval_s):
-                break
-
-    def _mark_metrics_failed(self, sid: str) -> bool:
-        with self._lock:
-            self._kv_usage[sid] = None
-            if sid not in self._metrics_fail_warned:
-                self._metrics_fail_warned.add(sid)
-                return True
-            return False
-
-    def _mark_metrics_success(self, sid: str, val: float | None) -> bool:
-        with self._lock:
-            self._kv_usage[sid] = val
-            if sid in self._metrics_fail_warned:
-                self._metrics_fail_warned.discard(sid)
-                return True
-            return False
-
-    def _kv_snapshot(self, server_ids: list[str]) -> dict[str, float | None]:
-        with self._lock:
-            return {sid: self._kv_usage.get(sid) for sid in server_ids}
-
-    def _refresh_metrics_blocking(self) -> None:
-        metric_name = self._metric_name
-        if not metric_name:
-            return
-
-        for sid in self._server_actor_ids:
-            url = build_metrics_url(sid, self._metrics_path)
-            try:
-                text = fetch_prometheus_text(url, self._fetch_timeout_s)
-                val = parse_prometheus_metric_value(text, metric_name)
-                if self._mark_metrics_success(sid, val):
-                    logger.info("KV cache metrics recovered for replica %s", sid)
-            except Exception as e:
-                if self._mark_metrics_failed(sid):
-                    logger.warning("Failed to refresh KV metrics for %s from %s: %s", sid, url, e)
-
-    def pick_server(
-        self,
-        server_ids: list[str],
-        inflight: dict[str, int],
-    ) -> str:
-        if self._metric_name:
-            usage = self._kv_snapshot(server_ids)
-        else:
-            usage = {sid: None for sid in server_ids}
-
-        def key(sid: str) -> tuple[float, int, str]:
-            kv = usage.get(sid)
-            if kv is None:
-                return (float("inf"), inflight[sid], sid)
-            return (kv, inflight[sid], sid)
-
-        return min(server_ids, key=key)
+    def acquire_server(self, request_id: str, request_group_id: str | None = None) -> str:
+        if request_group_id and request_group_id in self._request_group_to_server:
+            server_id = self._request_group_to_server[request_group_id]
+            self.update_inflight(server_id, 1)
+            return server_id
+        server_id = self.strategy.pick_server(list(self._server_actor_ids))
+        if request_group_id:
+            self._request_group_to_server[request_group_id] = server_id
+        self.update_inflight(server_id, 1)
+        return server_id
 
 
-@register("weighted_rr")
-class WeightedRoundRobinStrategy(LoadBalanceStrategy):
-    """Smooth weighted round-robin (largest-current-weight), static weights per replica order."""
-
-    def __init__(self, server_actor_ids: list[str], weights: list[float] | None = None, **kwargs: Any):
-        self._ids = list(server_actor_ids)
-        n = len(self._ids)
-        if n == 0:
-            raise ValueError("server_actor_ids must be non-empty")
-        if weights is None:
-            w = [1.0] * n
-        else:
-            if len(weights) != n:
-                raise ValueError(f"load_balance_weights length {len(weights)} != num servers {n}")
-            w = [float(x) for x in weights]
-        self._weights = w
-        self._current = [0.0] * n
-
-    def pick_server(
-        self,
-        server_ids: list[str],
-        inflight: dict[str, int],
-    ) -> str:
-        # Selection ignores inflight; weights define share across replicas.
-        pos = {sid: i for i, sid in enumerate(self._ids)}
-        indices = [pos[s] for s in server_ids]
-        total = sum(self._weights[i] for i in indices)
-        for i in indices:
-            self._current[i] += self._weights[i]
-        best_i = max(indices, key=lambda i: (self._current[i], -i))
-        self._current[best_i] -= total
-        return self._ids[best_i]
-
-
-@register("random")
-class RandomStrategy(LoadBalanceStrategy):
-    def __init__(self, server_actor_ids: list[str], random_seed: int | None = None, **kwargs: Any):
-        self._rng = random.Random(random_seed)
-
-    def pick_server(
-        self,
-        server_ids: list[str],
-        inflight: dict[str, int],
-    ) -> str:
-        return self._rng.choice(server_ids)
+_LOAD_BALANCER_ACTOR_CLASSES["request"] = RequestStickyLoadBalancer
+_LOAD_BALANCER_ACTOR_CLASSES["group"] = GroupStickyLoadBalancer
