+++ verl/experimental/fully_async_policy/agent_loop/agent_loop.py
@@ -53,7 +53,7 @@ class FullyAsyncLLMServerManager(AsyncLLMServerManager):
         sampling_params: dict[str, Any],
         image_data: Optional[list[Any]] = None,
         video_data: Optional[list[Any]] = None,
-        group_id: str | None = None,
+        request_group_id: str | None = None,
     ) -> TokenOutput:
         """Generate tokens from prompt ids.
 
@@ -91,7 +91,7 @@ class FullyAsyncLLMServerManager(AsyncLLMServerManager):
                 sampling_params=sampling_params,
                 image_data=image_data,
                 video_data=video_data,
-                group_id=group_id,
+                request_group_id=request_group_id,
             )
 
             # 2. merge output into final_output
@@ -144,10 +144,11 @@ class FullyAsyncAgentLoopWorker(AgentLoopWorker):
         teacher_servers: list[tuple[str, ray.actor.ActorHandle]] = None,
         teacher_load_balancer_handle: ray.actor.ActorHandle = None,
         reward_loop_worker_handles: list[ray.actor.ActorHandle] = None,
-        load_balance_group_id: str | None = None,
     ):
         self.server_manager = FullyAsyncLLMServerManager(
-            config, servers, load_balancer_handle, load_balance_group_id=load_balance_group_id
+            config,
+            servers,
+            load_balancer_handle,
         )
         super().__init__(
             config,
@@ -156,7 +157,6 @@ class FullyAsyncAgentLoopWorker(AgentLoopWorker):
             teacher_servers,
             teacher_load_balancer_handle,
             reward_loop_worker_handles,
-            load_balance_group_id=load_balance_group_id,
         )
