+++ verl/trainer/config/_generated_ppo_torchtitan_trainer.yaml
@@ -290,10 +290,9 @@ actor_rollout_ref:
       num_repeat_rollouts: null
     calculate_log_probs: false
     load_balance_strategy: least_requests
+    load_balance_sticky_mode: request
     load_balance_weights: null
     load_balance_random_seed: null
-    group_sticky_routing: false
-    num_load_balance_groups: 1
     kv_cache_metrics:
       _target_: verl.workers.config.KvCacheMetricsConfig
       refresh_interval_s: 2.0
